Fringe stopping (delay gen) change
Delay gens now generate channel indexes from -N/2 to N/2-1 instead
of 0 to N. This means that the phase offset and phase increment is
calculated for the centre of the band, and not from one edge as
previously. This makes the calculations more elegant as less info
(bandwidth) is needed. It also minimises the error in the centre of
the band and pushes the maximum errors to the edges of the band. This
change has been untested. No change in software is required.

Also fixed mask parameters for ct_tvg and pkt_cont_tvg that have the
number of channels as 13 instead of 10, likely as the they were copied
from narrow band.